Refactor so a card can have more than one face and print each face (support for double-sided cards)
